added get and post rooms in the backend
